Fix imports and logging. Move blacklist and priority abstractmethods to BaseMinerNeuron so that validators do not need to define them;
