+++ neurons/validator.py
@@ -23,7 +23,7 @@ import template
 from template.validator import forward
 
 # import base validator class which takes care of most of the boilerplate
-from template.base import BaseValidatorNeuron
+from template.base.validator import BaseValidatorNeuron
 
 
 class Neuron(BaseValidatorNeuron):
