--- template/base/__init__.py
@@ -1,2 +1,0 @@
-from .miner import BaseMinerNeuron
-from .validator import BaseValidatorNeuron
