+++ template/utils/__init__.py
@@ -1,4 +1,3 @@
 from . import config
 from . import misc
-from . import sync
 from . import uids
